+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/dao/AuthAuthorityDao.java
@@ -1,5 +1,6 @@
 package guru.qa.niffler.data.dao;
 
+import guru.qa.niffler.data.entity.auth.AuthUserEntity;
 import guru.qa.niffler.data.entity.auth.AuthorityEntity;
 
 import java.util.List;
@@ -16,6 +17,6 @@ public interface AuthAuthorityDao {
 
     Optional<AuthorityEntity> findByUserId(UUID userId);
 
-    void delete(AuthorityEntity authority);
+    void delete(AuthUserEntity authUser);
 
 }
