+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/UsersDbClient.java
@@ -1,6 +1,7 @@
 package guru.qa.niffler.service;
 
 import guru.qa.niffler.config.Config;
+import guru.qa.niffler.data.Databases;
 import guru.qa.niffler.data.Databases.XaFunction;
 import guru.qa.niffler.data.dao.impl.jdbc.AuthAuthorityDaoJdbc;
 import guru.qa.niffler.data.dao.impl.jdbc.AuthUserDaoJdbc;
@@ -23,82 +24,100 @@ import static guru.qa.niffler.data.Databases.xaTransaction;
 
 public class UsersDbClient {
 
-  private static final Config CFG = Config.getInstance();
-  private static final PasswordEncoder pe = PasswordEncoderFactories.createDelegatingPasswordEncoder();
+    private static final Config CFG = Config.getInstance();
+    private static final PasswordEncoder pe = PasswordEncoderFactories.createDelegatingPasswordEncoder();
 
 
-  public UserJson createUserSpringJdbc(UserJson user) {
-    AuthUserEntity authUser = new AuthUserEntity();
-    authUser.setUsername(user.username());
-    authUser.setPassword(pe.encode("12345"));
-    authUser.setEnabled(true);
-    authUser.setAccountNonExpired(true);
-    authUser.setAccountNonLocked(true);
-    authUser.setCredentialsNonExpired(true);
+    public UserJson createUserSpringJdbc(UserJson user) {
+        AuthUserEntity authUser = new AuthUserEntity();
+        authUser.setUsername(user.username());
+        authUser.setPassword(pe.encode("12345"));
+        authUser.setEnabled(true);
+        authUser.setAccountNonExpired(true);
+        authUser.setAccountNonLocked(true);
+        authUser.setCredentialsNonExpired(true);
 
-    AuthUserEntity createdAuthUser = new AuthUserDaoSpringJdbc(dataSource(CFG.authJdbcUrl()))
-        .create(authUser);
+        AuthUserEntity createdAuthUser = new AuthUserDaoSpringJdbc(dataSource(CFG.authJdbcUrl()))
+                .create(authUser);
 
-    AuthorityEntity[] authorityEntities = Arrays.stream(Authority.values()).map(
-        e -> {
-          AuthorityEntity ae = new AuthorityEntity();
-          ae.setUser(createdAuthUser);
-          ae.setAuthority(e);
-          return ae;
-        }
-    ).toArray(AuthorityEntity[]::new);
+        new AuthAuthorityDaoSpringJdbc(dataSource(CFG.authJdbcUrl()))
+                .create(createAuthorityArray(createdAuthUser));
 
-    new AuthAuthorityDaoSpringJdbc(dataSource(CFG.authJdbcUrl()))
-        .create(authorityEntities);
-
-    return UserJson.fromEntity(
-        new UdUserDaoSpringJdbc(dataSource(CFG.userdataJdbcUrl()))
-            .create(
-                    UdUserEntity.fromJson(user)
-            ),
-        null
-    );
-  }
+        return UserJson.fromEntity(
+                new UdUserDaoSpringJdbc(dataSource(CFG.userdataJdbcUrl()))
+                        .create(
+                                UdUserEntity.fromJson(user)
+                        ),
+                null
+        );
+    }
 
 
-  public UserJson createUser(UserJson user) {
-    return UserJson.fromEntity(
-        xaTransaction(
-            new XaFunction<>(
-                con -> {
-                  AuthUserEntity authUser = new AuthUserEntity();
-                  authUser.setUsername(user.username());
-                  authUser.setPassword(pe.encode("12345"));
-                  authUser.setEnabled(true);
-                  authUser.setAccountNonExpired(true);
-                  authUser.setAccountNonLocked(true);
-                  authUser.setCredentialsNonExpired(true);
-                  new AuthUserDaoJdbc(con).create(authUser);
-                  new AuthAuthorityDaoJdbc(con).create(
-                      Arrays.stream(Authority.values())
-                          .map(a -> {
-                                AuthorityEntity ae = new AuthorityEntity();
-                                ae.setUser(authUser);
-                                ae.setAuthority(a);
-                                return ae;
-                              }
-                          ).toArray(AuthorityEntity[]::new));
-                  return null;
-                },
-                CFG.authJdbcUrl()
-            ),
-            new XaFunction<>(
-                con -> {
-                    UdUserEntity ue = new UdUserEntity();
-                  ue.setUsername(user.username());
-                  ue.setFullname(user.fullname());
-                  ue.setCurrency(user.currency());
-                  new UdUserDaoJdbc(con).create(ue);
-                  return ue;
-                },
-                CFG.userdataJdbcUrl()
-            )
-        ),
-        null);
-  }
+    public UserJson createUser(UserJson user) {
+        return UserJson.fromEntity(
+                xaTransaction(
+                        new XaFunction<>(
+                                con -> {
+                                    AuthUserEntity authUser = new AuthUserEntity();
+                                    authUser.setUsername(user.username());
+                                    authUser.setPassword(pe.encode("12345"));
+                                    authUser.setEnabled(true);
+                                    authUser.setAccountNonExpired(true);
+                                    authUser.setAccountNonLocked(true);
+                                    authUser.setCredentialsNonExpired(true);
+                                    new AuthUserDaoJdbc(con).create(authUser);
+                                    new AuthAuthorityDaoJdbc(con).create(
+                                            Arrays.stream(Authority.values())
+                                                    .map(a -> {
+                                                                AuthorityEntity ae = new AuthorityEntity();
+                                                                ae.setUser(authUser);
+                                                                ae.setAuthority(a);
+                                                                return ae;
+                                                            }
+                                                    ).toArray(AuthorityEntity[]::new));
+                                    return null;
+                                },
+                                CFG.authJdbcUrl()
+                        ),
+                        new XaFunction<>(
+                                con -> {
+                                    UdUserEntity ue = new UdUserEntity();
+                                    ue.setUsername(user.username());
+                                    ue.setFullname(user.fullname());
+                                    ue.setCurrency(user.currency());
+                                    new UdUserDaoJdbc(con).create(ue);
+                                    return ue;
+                                },
+                                CFG.userdataJdbcUrl()
+                        )
+                ),
+                null);
+    }
+
+    public void deleteUser(UserJson user) {
+        xaTransaction(new Databases.XaConsumer(connection -> new AuthUserDaoJdbc(connection)
+                        .findByUsername(user.username())
+                        .ifPresent(
+                                authUser -> {
+                                    new AuthAuthorityDaoJdbc(connection).delete(authUser);
+                                    new AuthUserDaoJdbc(connection).delete(authUser);
+                                }
+                        ), CFG.authJdbcUrl()),
+                new Databases.XaConsumer(connection -> new UdUserDaoJdbc(connection)
+                        .findByUsername(user.username())
+                        .ifPresent(
+                                ue -> new UdUserDaoJdbc(connection).delete(ue)
+                        ), CFG.userdataJdbcUrl()));
+    }
+
+    private AuthorityEntity[] createAuthorityArray(AuthUserEntity authUser) {
+        return Arrays.stream(Authority.values())
+                .map(authority -> {
+                            AuthorityEntity ae = new AuthorityEntity();
+                            ae.setUser(authUser);
+                            ae.setAuthority(authority);
+                            return ae;
+                        }
+                ).toArray(AuthorityEntity[]::new);
+    }
 }
